[FIX] fields.py: provide fields a default name
From python code, import a module class and print one of its field, it
creates an infinite recursion due to `model_name` and `name` field
attributes being accessed by `__str__` but only being set when
accessing a model through the environment.

This commit provide a default values for the field name and the field
model name to prevent that infinite recursion.